Put cocci SMPL into separate files
The cocci semantic patches were stored inside the code and this was
not convenient for modification. This patch move every SMPL into
a separate file and use a simple template system to improve the
process of building the temporary cocci file.

Due to this data introduction, the setup.py has evolved to incorporate
this new data files. This is result and aan important source tree
modification.